+++ ios/DiscoverSheet/UIViewController+PanModalPresenter.swift
@@ -4,8 +4,8 @@ import PanModal
 
 class BetterGestureRecognizerDelegateAdapter: NSObject, UIGestureRecognizerDelegate {
   var grd: UIGestureRecognizerDelegate
-  var config: NSObject
-  required init(grd: UIGestureRecognizerDelegate, config: NSObject) {
+  var config: InvisibleView
+  required init(grd: UIGestureRecognizerDelegate, config: InvisibleView) {
     self.grd = grd
     self.config = config
     super.init()
@@ -13,84 +13,81 @@ class BetterGestureRecognizerDelegateAdapter: NSObject, UIGestureRecognizerDeleg
   public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldBeRequiredToFailBy otherGestureRecognizer: UIGestureRecognizer) -> Bool {
     return self.grd.gestureRecognizer!(gestureRecognizer, shouldBeRequiredToFailBy: otherGestureRecognizer)
   }
-
+  
   public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
     return self.grd.gestureRecognizer!(gestureRecognizer, shouldRecognizeSimultaneouslyWith: otherGestureRecognizer)
   }
-
+  
   public func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRequireFailureOf otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-
+    
     if (String(describing: type(of: otherGestureRecognizer))
-      == "UIScrollViewPanGestureRecognizer"
-      && (grd as? UIPresentationController)?.presentedViewController is PanModalPresentable && otherGestureRecognizer.view != ((grd as? UIPresentationController)?.presentedViewController as? PanModalPresentable)?.panScrollable) {
-      return self.config.value(forKey: "interactsWithOuterScrollView") as! Bool
+          == "UIScrollViewPanGestureRecognizer"
+          && (grd as? UIPresentationController)?.presentedViewController is PanModalPresentable && otherGestureRecognizer.view != ((grd as? UIPresentationController)?.presentedViewController as? PanModalPresentable)?.panScrollable) {
+      return self.config.interactsWithOuterScrollView
     }
     return false
   }
 }
 
-var moved = false
 
-class PossiblyTouchesPassableUIView: UIView {
-  var grdelegate: UIGestureRecognizerDelegate?
-  var config: NSObject?
-  var topLayoutGuideLength: CGFloat?
-  var oldClass: AnyClass?
-  var justModifiedClass: NSNumber?
-
-
-  override func hitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
-    let blocksBackgroundTocuhes = self.config?.value(forKey: "blocksBackgroundTouches") as! Bool
-    if (blocksBackgroundTocuhes || self.subviews[1].frame.contains(point)) {
-      return super.hitTest(point, with: event)
-    }
-    return nil
+extension Collection {
+  subscript (safe index: Index) -> Element? {
+    return indices.contains(index) ? self[index] : nil
   }
-  //  // I don't really want to talk about it
-  override func layoutSubviews() {
-    super.layoutSubviews()
+}
 
-    let outerView = self.config?.value(forKey: "outerView") as? UIView
-    if (!(self.config!.value(forKey: "presentGlobally") as! Bool)) {
-      if moved {
-        return
+extension UIView {
+  
+  func getHelperView() -> HelperView? {
+    if self.subviews.count > 1 &&
+        self.subviews[1].subviews.count > 0 &&
+        self.subviews[1].subviews[0] is HelperView &&
+        (self.subviews[1].subviews[0] as! HelperView).controller != nil {
+      return (self.subviews[1].subviews[0] as! HelperView);
+    }
+    return nil;
+  }
+  
+  @objc func betterHitTest(_ point: CGPoint, with event: UIEvent?) -> UIView? {
+    let helperView = getHelperView()
+    if helperView != nil {
+      let config = helperView!.controller!.config
+      let blocksBackgroundTocuhes = config?.blocksBackgroundTouches;
+      if (blocksBackgroundTocuhes! || self.subviews[1].frame.contains(point)) {
+        return self.betterHitTest(point, with: event)
       }
-      moved = true
-      removeFromSuperview()
-      let helperView: UIView = self.subviews[1].subviews[0]
+      return nil
+    }
+    return self.betterHitTest(point, with: event)
+  }
+  
+  @objc func betterLayoutSubviews() {
+    self.betterLayoutSubviews()
+    let helperView = getHelperView()
+    if helperView != nil {
+      
+      let controller = helperView!.controller!
+      let config = controller.config
+      
+      let outerView = config?.value(forKey: "outerView") as? UIView
+      if (!(config!.presentGlobally)) {
+        if controller.moved {
+          return
+        }
+        controller.moved = true
+        removeFromSuperview()
         let bounds = outerView!.bounds
-        let topOffset:CGFloat = (topLayoutGuideLength ?? 0) + CGFloat(truncating: self.config?.value(forKey: "topOffset") as! NSNumber)
-
+        let topOffset:CGFloat = (controller.topLayoutGuide.length ?? 0) + CGFloat(truncating: config?.topOffset as! NSNumber)
         let newBounds = CGRect.init(x: bounds.minX, y: bounds.minY, width: bounds.width, height: bounds.height - topOffset)
-        helperView.setValue(newBounds, forKeyPath: "specialBounds")
-      outerView?.addSubview(self)
-    }
-    let gr: UIGestureRecognizer = self.gestureRecognizers![0]
-    grdelegate = BetterGestureRecognizerDelegateAdapter.init(grd: gr.delegate!, config: config!)
-    gr.delegate = grdelegate
-    if outerView == nil {
-      makeOldClass()
-    }
-  }
-
-  @objc func makeOldClass() {
-    if self.oldClass != nil {
-      let oldClassMem: AnyClass = self.oldClass!
-      self.oldClass = nil
-      object_setClass(self, oldClassMem)
-    }
-  }
-
-  override func didMoveToWindow() {
-    if self.window == nil {
-      if (justModifiedClass?.isEqual(to: true))! {
-        justModifiedClass = NSNumber.init(value: false)
-        super.didMoveToWindow()
-        return
+        helperView!.setSpecialBounds(newBounds)
+        outerView?.addSubview(self)
+        
+        let gr: UIGestureRecognizer = self.gestureRecognizers![0]
+        controller.grdelegate = BetterGestureRecognizerDelegateAdapter.init(grd: gr.delegate!, config: config!)
+        gr.delegate = controller.grdelegate
       }
-      makeOldClass()
     }
-    super.didMoveToWindow()
+    
   }
 }
 
@@ -98,15 +95,15 @@ var PossiblyTouchesPassableUITransitionView: AnyClass?  = nil;
 
 class DiscoverSheetViewController: UIViewController, PanModalPresentable {
   func hide() {
-
+    
   }
-
+  
   func unhackParent() {
-
+    
   }
-
+  
   var observation: NSKeyValueObservation?
-
+  
   @objc func jumpTo(long: NSNumber) {
     self.panModalSetNeedsLayoutUpdate()
     if (long.boolValue) {
@@ -116,27 +113,25 @@ class DiscoverSheetViewController: UIViewController, PanModalPresentable {
     }
     self.panModalSetNeedsLayoutUpdate()
   }
-
-  var config: NSObject?
+  
+  var config: InvisibleView?
+  var grdelegate: UIGestureRecognizerDelegate?
+  var topLayoutGuideLength: CGFloat?
+  var moved = false;
   var scrollView: UIScrollView?
-  convenience init(config: NSObject) {
+  convenience init(config: InvisibleView) {
     self.init()
     self.config = config
   }
-
-  override func viewDidLoad() {
-    super.viewDidLoad()
-    hack()
-  }
-
+  
   @objc func panModalSetNeedsLayoutUpdateWrapper() {
     panModalSetNeedsLayoutUpdate()
   }
-
+  
   override var preferredStatusBarStyle: UIStatusBarStyle {
     return .lightContent
   }
-
+  
   func findChildScrollViewDFS(view: UIView)-> UIScrollView? {
     var viewsToTraverse = [view]
     while !viewsToTraverse.isEmpty {
@@ -144,7 +139,7 @@ class DiscoverSheetViewController: UIViewController, PanModalPresentable {
       viewsToTraverse.removeLast()
       if last is UIScrollView {
         scrollView = (last as! UIScrollView);
-        scrollView!.scrollsToTop = self.config?.value(forKey: "scrollsToTop") as! Bool
+        scrollView!.scrollsToTop = self.config?.scrollsToTop as! Bool
         return last as? UIScrollView
       }
       last.subviews.forEach { subview in
@@ -153,138 +148,129 @@ class DiscoverSheetViewController: UIViewController, PanModalPresentable {
     }
     return nil
   }
-
+  
   @objc func setScrollsToTop(scrollsToTop: NSNumber) {
     scrollView?.scrollsToTop = scrollsToTop.boolValue;
   }
-
+  
   var panScrollable: UIScrollView? {
     return findChildScrollViewDFS(view: self.view!)
   }
-
+  
   var shortFormHeight: PanModalHeight {
-    let height: CGFloat = CGFloat(truncating: self.config?.value(forKey: "shortFormHeight") as! NSNumber)
+    let height: CGFloat = CGFloat(truncating: self.config?.shortFormHeight ?? 0)
     return isShortFormEnabled ? .contentHeight(height) : longFormHeight
   }
-
+  
   var topOffset: CGFloat {
-    let topOffset: CGFloat = CGFloat(truncating: self.config?.value(forKey: "topOffset") as! NSNumber)
+    let topOffset: CGFloat = CGFloat(truncating: self.config?.topOffset ?? 0)
     return topLayoutGuide.length + topOffset
   }
-
+  
   var isShortFormEnabledInternal = 2
   var isShortFormEnabled: Bool {
-    let startFromShortForm = self.config?.value(forKey: "startFromShortForm") as! Bool
+    let startFromShortForm = self.config?.startFromShortForm ?? false
     if isShortFormEnabledInternal > 0 && !startFromShortForm {
       isShortFormEnabledInternal -= 1
       return false
     }
-    return self.config?.value(forKey: "isShortFormEnabled") as! Bool
+    return self.config?.isShortFormEnabled ?? false
   }
-
+  
   var longFormHeight: PanModalHeight {
-    if self.config?.value(forKey: "longFormHeight") == nil {
+    if self.config?.longFormHeight == nil {
       return .maxHeight
     }
-    return .contentHeight(CGFloat(truncating: self.config?.value(forKey: "longFormHeight") as! NSNumber))
+    return .contentHeight(CGFloat(truncating: self.config?.longFormHeight ?? 0))
   }
-
+  
   var cornerRadius: CGFloat {
-    return CGFloat(truncating: self.config?.value(forKey: "cornerRadius") as! NSNumber)
+    return CGFloat(truncating: self.config?.cornerRadius ?? 0)
   }
-
+  
   var springDamping: CGFloat {
-    return CGFloat(truncating: self.config?.value(forKey: "springDamping") as! NSNumber)
+    return CGFloat(truncating: self.config?.springDamping ?? 0)
   }
-
+  
   var isInitialAnimation = 3
-
+  
   var transitionDuration: Double {
-    if isInitialAnimation > 0 && !(self.config?.value(forKey: "initialAnimation") as! Bool)
+    if isInitialAnimation > 0 && !(self.config?.initialAnimation ?? false)
     {
       isInitialAnimation -= 1
       return 0.0
     }
-    return Double(truncating: self.config?.value(forKey: "transitionDuration") as! NSNumber)
+    return Double(truncating: self.config?.transitionDuration ?? 0)
   }
-
+  
   var panModalBackgroundColor: UIColor {
-    return UIColor.black.withAlphaComponent(CGFloat(truncating: self.config?.value(forKey: "backgroundOpacity") as! NSNumber))
-
+    return UIColor.black.withAlphaComponent(CGFloat(truncating: self.config?.backgroundOpacity ?? 0))
+    
   }
   var anchorModalToLongForm: Bool {
-    return self.config?.value(forKey: "anchorModalToLongForm") as! Bool
+    return self.config?.anchorModalToLongForm ?? false
   }
-
+  
   var allowsDragToDismiss: Bool {
-    return self.config?.value(forKey: "allowsDragToDismiss") as! Bool
+    return self.config?.allowsDragToDismiss ?? false
   }
-
+  
   var allowsTapToDismiss: Bool {
-    return self.config?.value(forKey: "allowsTapToDismiss") as! Bool
+    return self.config?.allowsTapToDismiss ?? false
   }
-
+  
   var isUserInteractionEnabled: Bool {
-    return self.config?.value(forKey: "isUserInteractionEnabled") as! Bool  }
-
+    return self.config?.isUserInteractionEnabled ?? false  }
+  
   var isHapticFeedbackEnabled: Bool {
-    return self.config?.value(forKey: "isHapticFeedbackEnabled") as! Bool  }
-
+    return self.config?.isHapticFeedbackEnabled ?? false  }
+  
   var shouldRoundTopCorners: Bool {
     hack()
-    return self.config?.value(forKey: "shouldRoundTopCorners") as! Bool
+    return self.config?.shouldRoundTopCorners ?? false
   }
-
+  
   override func viewDidLayoutSubviews() {
     hack()
   }
-
+  
   func updatePostion(position: CGFloat, newPostion: CGFloat) {
     if (position < 80.0 && newPostion >= 80.0) {
       self.config?.performSelector(inBackground: Selector.init(("callOnCrossMagicBoderFromBottom")), with: nil)
     }
-
+    
     if (position >= 80.0 && newPostion < 80.0) {
       self.config?.performSelector(inBackground: Selector.init(("callOnCrossMagicBoderFromTop")), with: nil)
     }
   }
-
+  
   var hacked = false
-
+  
   public func hack() {
-    let pview = view.superview!.superview!
-    if !(pview is PossiblyTouchesPassableUIView && !hacked) {
-      let oldClass: AnyClass = type(of: pview)
-      object_setClass(pview, PossiblyTouchesPassableUIView.self)
-      hacked = true;
-      (pview as! PossiblyTouchesPassableUIView).config = self.config
-      (pview as! PossiblyTouchesPassableUIView).topLayoutGuideLength = self.topLayoutGuide.length
-      (pview as! PossiblyTouchesPassableUIView).oldClass = oldClass
-      (pview as! PossiblyTouchesPassableUIView).justModifiedClass = NSNumber.init(value: true)
+    if !hacked {
       observation = (self.presentationController as! PanModalPresentationController).observe(\.yPosition, options: [.old, .new]) { object, change in
-        self.updatePostion(position: change.oldValue!, newPostion: change.newValue!)
-      }
+        self.updatePostion(position: change.oldValue!, newPostion: change.newValue!) }
     }
   }
-
+  
   var showDragIndicator: Bool {
-    return self.config?.value(forKey: "showDragIndicator") as! Bool
+    return self.config?.showDragIndicator ?? false
   }
-
+  
   var scrollIndicatorInsets: UIEdgeInsets {
     let bottomOffset = presentingViewController?.bottomLayoutGuide.length ?? 0
     return UIEdgeInsets(top: 0, left: 0, bottom: bottomOffset, right: 0)
   }
-
-
+  
+  
   func shouldPrioritize(panModalGestureRecognizer: UIPanGestureRecognizer) -> Bool {
-    let headerHeight: CGFloat = CGFloat(truncating: self.config?.value(forKey: "headerHeight") as! NSNumber)
+    let headerHeight: CGFloat = CGFloat(truncating: self.config?.headerHeight ?? 0)
     let location = panModalGestureRecognizer.location(in: view)
     return location.y < headerHeight
   }
-
+  
   func willTransition(to state: PanModalPresentationController.PresentationState) {
-    if self.config?.value(forKey: "onWillTransition") != nil {
+    if self.config?.onWillTransition != nil {
       if state == .longForm {
         self.config?.performSelector(inBackground: Selector.init(("callWillTransitionLong")), with: nil)
       } else {
@@ -292,25 +278,42 @@ class DiscoverSheetViewController: UIViewController, PanModalPresentable {
       }
     }
   }
-
+  
   func panModalWillDismiss() {
-    if self.config?.value(forKey: "onWillDismiss") != nil {
+    if self.config?.onWillDismiss != nil {
       self.config?.performSelector(inBackground: Selector.init(("callWillDismiss")), with: nil)
     }
   }
-
+  
   func panModalDidDismiss() {
-    if self.config?.value(forKey: "onDidDismiss") != nil {
+    if self.config?.onDidDismiss != nil {
       self.config?.performSelector(inBackground: Selector.init(("callDidDismiss")), with: nil)
     }
   }
 }
 
+var swizzled = false;
+
+func swizzle(uiTransitionView: UIView?) {
+  if uiTransitionView == nil {
+    return;
+  }
+  let UITransitionView: AnyClass = type(of: uiTransitionView!)
+  if !swizzled {
+    swizzled = true;
+    let originalMethodLS = class_getInstanceMethod(UITransitionView, #selector(UIView.layoutSubviews))
+    let swizzledMethodLS = class_getInstanceMethod(UITransitionView, #selector(UIView.betterLayoutSubviews))
+    method_exchangeImplementations(originalMethodLS!, swizzledMethodLS!)
+    let originalMethodHT = class_getInstanceMethod(UITransitionView, #selector(UIView.hitTest(_:with:)))
+    let swizzledMethodHT = class_getInstanceMethod(UITransitionView, #selector(UIView.betterHitTest(_:with:)))
+    method_exchangeImplementations(originalMethodHT!, swizzledMethodHT!)
+  }
+}
 
 extension UIViewController {
-  @objc public func presentPanModal(view: UIView, config: UIView) -> UIViewController? {
-    moved = false;
-    if self.presentedViewController != nil {
+  @objc public func presentPanModal(view: HelperView, config: InvisibleView) {
+    if self.presentedViewController != nil || !swizzled {
+      swizzle(uiTransitionView: config.window?.rootViewController?.view.superview?.superview)
       DispatchQueue.main.asyncAfter(deadline: .now() + 0.03) {
         self.presentPanModal(view: view, config: config)
       }
@@ -319,8 +322,8 @@ extension UIViewController {
       viewControllerToPresent.view = view
       let sourceView: UIView? = nil, sourceRect: CGRect = .zero
       self.presentPanModal(viewControllerToPresent, sourceView: sourceView, sourceRect: sourceRect)
-      return viewControllerToPresent
+      view.controller = (viewControllerToPresent as! DiscoverSheetViewController);
+      config.contoller = (viewControllerToPresent as! DiscoverSheetViewController);
     }
-    return nil
   }
 }